backend/requirements.txt
fastapi>=0.95.0,<0.100.0
uvicorn[standard]>=0.22.0,<0.23.0
pydantic>=1.10.0,<2.0.0
email-validator>=1.3.0,<2.0.0
python-multipart>=0.0.5,<0.0.7
sqlalchemy>=1.4.0,<2.0.0
passlib[bcrypt]>=1.7.4,<1.8.0
python-jose[cryptography]>=3.3.0,<3.4.0
psycopg2-binary>=2.9.0,<3.0.0
cloudinary>=1.33.0,<2.0.0
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
gunicorn>=20.1.0,<21.0.0